Change generic to xbmc 12+ standard
